// src/model/dummyjson/IDJUser.ts
export interface ICoordinates {
    lat: number;
    lng: number;
}

export interface IAddress {
    address: string;
    city: string;
    state: string;
    stateCode: string;
    postalCode: string;
    coordinates: ICoordinates;
    country: string;
}

export interface IHair {
    color: string;
    type: string;
}

export interface ICompany {
    department: string;
    name: string;
    title: string;
    address: IAddress;
}

export interface IDJUser {
    id: number;
    firstName: string;
    lastName: string;
    maidenName: string;
    age: number;
    gender: string;
    email: string;
    phone: string;
    username: string;
    birthDate: string;
    image: string;
    bloodGroup: string;
    height: number;
    weight: number;
    eyeColor: string;
    hair: IHair;
    ip: string;
    address: IAddress;
    macAddress: string;
    university: string;
    company: ICompany;
    role: string;
}